Fix compilation error when not encrypting split encryption to some modules for convenience
