+++ crates/smb/Cargo.toml
@@ -45,7 +45,7 @@ byteorder = { version = "1.5.0", optional = true }
 sspi = { version = "0.18.0", features = ["ring"], default-features = false }
 reqwest = { workspace = true, optional = true }
 
-# Crypto
+# Crypto; RustCrypto provides support for RC versions only.
 hmac = "0.13.0-rc.2"
 sha2 = "0.11.0-rc.2"
 rust-kbkdf = { version = "1.1.1" }
@@ -96,11 +96,12 @@ compress_lz4 = ["dep:lz4_flex"]
 compress = ["compress_pattern_v1", "compress_lz4"]
 
 # Encryption
-encrypt_aes128ccm = ["dep:ccm", "dep:aead"]
-encrypt_aes256ccm = ["dep:ccm", "dep:aead"]
+__encrypt_core = []
+encrypt_aes128ccm = ["dep:ccm", "dep:aead", "__encrypt_core"]
+encrypt_aes256ccm = ["dep:ccm", "dep:aead", "__encrypt_core"]
+encrypt_aes128gcm = ["dep:aes-gcm", "dep:aead", "__encrypt_core"]
+encrypt_aes256gcm = ["dep:aes-gcm", "dep:aead", "__encrypt_core"]
 encrypt_aesccm = ["encrypt_aes128ccm", "encrypt_aes256ccm"]
-encrypt_aes128gcm = ["dep:aes-gcm", "dep:aead"]
-encrypt_aes256gcm = ["dep:aes-gcm", "dep:aead"]
 encrypt_aesgcm = ["encrypt_aes128gcm", "encrypt_aes256gcm"]
 encrypt = ["encrypt_aesccm", "encrypt_aesgcm"]
